requirements.txt
pynput>=1.7.6
pyautogui>=0.9.54
pyperclip>=1.8.2
easygui>=0.98.3
Pillow>=9.0.0